Move code around for readability, change boolean to enum
