+++ src/ServiceControl.Persistence.EFCore/Entities/ArchiveOperationType.cs
@@ -1,0 +1,7 @@
+namespace ServiceControl.Persistence.EFCore.Entities;
+
+public enum ArchiveOperationType
+{
+    Archive,
+    UnArchive
+}
