+++ src/ServiceControl.Persistence.EFCore/EntityConfigurations/ArchiveOperationConfiguration.cs
@@ -8,16 +8,14 @@ class ArchiveOperationConfiguration : IEntityTypeConfiguration<ArchiveOperationE
 {
     public void Configure(EntityTypeBuilder<ArchiveOperationEntity> builder)
     {
-        builder.ToTable("ArchiveOperations");
-
         // Composite primary key — the natural key that distinguishes one operation from another.
-        // This also serves as the uniqueness constraint: one operation per (RequestId, ArchiveType, IsArchive).
-        builder.HasKey(e => new { e.RequestId, e.ArchiveType, e.IsArchive });
+        // This also serves as the uniqueness constraint: one operation per (RequestId, ArchiveType, OperationType).
+        builder.HasKey(e => new { e.RequestId, e.ArchiveType, e.OperationType });
 
         builder.Property(e => e.RequestId).HasMaxLength(64).IsRequired();
         builder.Property(e => e.GroupName).IsRequired();
         builder.Property(e => e.ArchiveType).IsRequired();
-        builder.Property(e => e.IsArchive).IsRequired();
+        builder.Property(e => e.OperationType).IsRequired();
         builder.Property(e => e.TotalNumberOfMessages).IsRequired();
         builder.Property(e => e.NumberOfMessagesProcessed).IsRequired();
         builder.Property(e => e.NumberOfBatches).IsRequired();
